Capistrano question and Readme

+++ template.rb
@@ -34,10 +34,19 @@ def use_doorkeeper?
   @doorkeeper
 end
 
+def ask_capistrano
+  @capistrano = ask("Do you want to use Capistrano? (y / n)") == "y" ? true : false
+end
+
+def use_capistrano?
+  @capistrano
+end
+
 def apply_template!
   assert_minimum_rails_version
   assert_postgresql
   ask_doorkeeper
+  ask_capistrano
   add_template_repository_to_source_path
 end
 
@@ -50,17 +59,14 @@ def add_dependencies
     gem 'faker'
   end
 
-  gem_group :development do
-    gem 'web-console', '>= 3.3.0'
-    gem 'listen', '>= 3.0.5', '< 3.2'
-    gem 'spring'
-    gem 'spring-watcher-listen', '~> 2.0.0'
-  
-    gem 'capistrano'
-    gem 'capistrano-rails'
-    gem 'capistrano3-unicorn'
-    gem 'capistrano-rvm'
-    gem 'capistrano-unicorn-monit', github: 'bypotatoes/capistrano-unicorn-monit'
+  if use_capistrano?
+    gem_group :development do
+      gem 'capistrano'
+      gem 'capistrano-rails'
+      gem 'capistrano3-unicorn'
+      gem 'capistrano-rvm'
+      gem 'capistrano-unicorn-monit', github: 'bypotatoes/capistrano-unicorn-monit'
+    end
   end
   
   gem_group :test do
@@ -122,7 +128,6 @@ def prepare_doorkeeper
 
     gsub_file "config/initializers/wine_bouncer.rb", "config.auth_strategy = :default", "config.auth_strategy = :swagger"
   end
-  run "bundle exec cap install"
 end
 
 def prepare_environment
@@ -233,15 +238,29 @@ def copy_initializers
   run 'cp -r lib/exi-monolith/config/initializers config/initializers'
 end
 
+def prepare_capistrano
+  run "bundle exec cap install" if use_capistrano?
+end
+
 def setup_capistrano
-  run 'rm Capfile'
-  run 'cp -r lib/exi-monolith/Capfile .'
+  if use_capistrano?
+    # Capfile
+    gsub_file 'Capfile', '# require "capistrano/rvm"', 'require "capistrano/rvm"'
+    gsub_file 'Capfile', '# require "capistrano/bundler"', 'require "capistrano/bundler"'
+    gsub_file 'Capfile', '# require "capistrano/rails/assets"', 'require "capistrano/rails/assets"'
+    gsub_file 'Capfile', '# require "capistrano/rails/migrations"', 'require "capistrano/rails/migrations"'
+    insert_into_file "Capfile", "\nrequire 'capistrano/seed_migration_tasks'", after: '# require "capistrano/passenger"'
+    insert_into_file "Capfile", "\nrequire 'capistrano/unicorn'", after: "require 'capistrano/seed_migration_tasks'"
+    insert_into_file "Capfile", "\nrequire 'capistrano/unicorn/monit'", after: "require 'capistrano/unicorn'"
 
-  run 'rm -r config/deploy.rb'
-  run 'cp -r lib/exi-monolith/config/deploy.rb config/deploy.rb'
-
-  run 'rm -r config/deploy'
-  run 'cp -r lib/exi-monolith/config/deploy config/deploy'
+    # deploy.rb
+    gsub_file "config/deploy.rb", '# append :linked_dirs, "log", "tmp/pids", "tmp/cache", "tmp/sockets", "public/system"', 'append :linked_dirs, "log", "tmp/pids", "tmp/cache", "tmp/sockets", "public/system"'
+    insert_into_file "config/deploy.rb", "\n\nafter 'deploy:migrating', 'seed:migrate'", after: "# set :ssh_options, verify_host_key: :secure"
+    
+    # config/deploy/
+    run 'rm -r config/deploy'
+    run 'cp -r lib/exi-monolith/config/deploy config/deploy'
+  end
 end
 
 def stop_spring
@@ -294,6 +313,7 @@ after_bundle do
   override_database_yml
   prepare_environment
   copy_initializers
+  prepare_capistrano
   setup_capistrano
   finishing
   stop_spring
